bit of hookage
